requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-on-heroku==1.1.2
django-session-timeout==0.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
six==1.16.0
sqlparse==0.4.2
whitenoise==6.2.0
